requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
Django==2.2.6
django-crispy-forms==1.7.2
django-multiselectfield==0.1.9
geographiclib==1.50
geojson==2.5.0
geopy==1.20.0
idna==2.8
Pillow==6.2.1
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
stripe==2.40.0
urllib3==1.25.6
